fix: node file structure, colcon build directories
