+++ src/ug_sensors/createpkg.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+ros2 pkg create <pkg-name> --dependencies [deps] --build-type ament_python
