+++ src/ug_sensors/setup.cfg
