+++ src/ug_sensors/setup.py
@@ -1,0 +1,23 @@
+from setuptools import find_packages, setup
+
+package_name = "sensor"
+
+setup(
+    name=package_name,
+    version="0.0.0",
+    packages=find_packages(exclude=["test"]),
+    data_files=[
+        ("share/ament_index/resource_index/packages", ["resource/" + package_name]),
+        ("share/" + package_name, ["package.xml"]),
+    ],
+    install_requires=["setuptools"],
+    zip_safe=True,
+    maintainer="Anthony Cieri",
+    maintainer_email="penguinmillion@gmail.com",
+    description="TODO: Package description",
+    license="MIT",
+    tests_require=["pytest"],
+    entry_points={
+        "console_scripts": ["baro_node=sensor.depth_sensor:main"],
+    },
+)
